Fix landscape layout issue in celebration/settings cells.
We create two versions for these cells: a portrait and a landscape
version.

This is a hack, but for celebration cells this might be the only viable
solution: there is no way (at least in IB) to set the label width to
cellWidth-<delta>, only to contentViewWidth-<delta>. But the size of
contentView depends on whether the checkbox accessory is enabled or not.
So it is better if we set the width of the label directly - but we have
to set it differently for portrait and landscape mode.